third_party/blink/web_tests/http/tests/test-runner/storage-isolation-1-expected.txt

CONSOLE MESSAGE: --- [storage-isolation-1.html] Starting Test ---
CONSOLE MESSAGE: Verifying initial state...
CONSOLE MESSAGE:   OK: LocalStorage is empty.
CONSOLE MESSAGE:   OK: IndexedDB is empty.
CONSOLE MESSAGE:   OK: CacheStorage is empty.
CONSOLE MESSAGE:   OK: Origin Private File System is empty.
CONSOLE MESSAGE: Creating data...
CONSOLE MESSAGE: --- [storage-isolation-1.html] Test Finished ---
